Add new registration page for user at login page
Add registration page at login page, but meet some problem that the former version had already use auth function trying to add register function, which is conflit with the "form solution" I use. There is setting in the setting.py, which hard to change and also cause new account cannot be generated. Need to enable the "register" later with solving the conflict.